pyproject.toml
[project]
name = "tgfeed"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.42.0",
    "python-dotenv>=1.0.0",
    "bottle>=0.12.0",
    "waitress>=3.0.0",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.11",
]
